requirements.txt
asgiref==3.5.1
attrs==21.4.0
autobahn==22.4.2
Automat==20.2.0
certifi==2021.10.8
cffi==1.15.0
channels==3.0.4
charset-normalizer==2.0.12
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==37.0.2
daphne==3.0.2
Django==4.0.4
django-filter==21.1
django-sendsms==0.5
djangorestframework==3.13.1
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.1.0
drf-yasg==1.20.0
drfaddons==0.1.4
hyperlink==21.0.0
idna==3.3
incremental==21.3.0
inflection==0.5.1
iniconfig==1.1.1
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.1
model-bakery==1.5.0
packaging==21.3
Pillow==9.1.0
pluggy==1.0.0
py==1.11.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyJWT==1.7.1
pyOpenSSL==22.0.0
pyparsing==3.0.9
pytest==7.1.2
python-decouple==3.6
pytz==2022.1
requests==2.27.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
service-identity==21.1.0
six==1.16.0
sqlparse==0.4.2
tomli==2.0.1
Twisted==22.4.0
txaio==22.2.1
typing_extensions==4.2.0
uritemplate==4.1.1
urllib3==1.26.9
zope.interface==5.4.0
